testing watch event
Was able to get grunt-contrib-copy to properly read 1 file at a time.
LESS/CSSMIN giving problems, most likely due to the very specific
handling of said files. Probably faster to just process them as is.
There is only 4 files total right now so it wouldn't cause much of a
problem.